third_party/WebKit/LayoutTests/http/tests/websocket/interleaved-fragments-expected.txt

CONSOLE ERROR: line 15: WebSocket connection to 'ws://localhost:8880/interleaved-fragments' failed: Received start of new message but previous message is unfinished.
Test whether WebSocket rejects interleaved fragmented frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
